requirements.txt
python-dotenv
langchain-groq
langchain-google-firestore
